% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleID.R
\name{get_user_info}
\alias{get_user_info}
\title{Get the logged in user's email and other info}
\usage{
get_user_info(id = "me")
}
\arguments{
\item{id}{ID of the person to get the profile data for. 'me' to get current user.}
}
\value{
A People resource

https://developers.google.com/+/web/api/rest/latest/people#resource-representations
}
\description{
Get the logged in user's email and other info
}

